Move lifecycle tests to lifecycle_test.go
